Remove alias_method_chain for render & render_to_string
Before, you would invoke wicked_pdf by calling:

    render :pdf => 'mypdf'

But this hooks into *every* render call.
The new way to do this would be to call `wicked_pdf` from a controller like:

    wicked_pdf :pdf => 'mypdf'

And since we don't need to check for the `:pdf` key any more, it is
now optional, and the method name will be used if one isn't supplied.

    wicked_pdf # same thing